ENGR00233866: SAHARA4 RNG fails to auto-seed
Problem description:
The SAHARA driver hangs when more than 8192x1024 bytes of random data
are generated.

Explanation:
When auto-seed is enabled, SAHARA generates a seed every 4094x1024bytes.
The first seed is generated when auto-seed is set during the
initialization. At the same time, a second seed is generated, and will
replace the first one when 4094x1024bytes of random data are generated.
When it reaches that required number of bytes, a third seed should be
generated in anticipation to replace the second seed.
That seed is never generated, and the status bit request for
seed is set.
The next time the number of generated bytes reaches 4094x1024, the
descriptor that requests random data cannot be processed.
The driver hangs there waiting for the descriptor to finish, and this
never occurs. A reset of SAHARA does not recover from there, and a reset
of the chip is necessary.

* Tried board:
i.MX51 EVK
This shouldn't be related to a board, just need a chip with SAHARA4 such
i.MX51.
i.MX53 might be affected too, but I did not verify it.

* Detailed reproduce steps:
(1) Replace the attached rand.c in place of the default in the imx-test
package. It requests more than 8192x1024bytes which is the limit when it
hangs. It displays the amount of generated bytes every 4096x1024bytes.
(2) Use the attached 'go' script to only run the random data test in the
unit_tests.
The test will hang.

Solution:
After discussing with STC design team, there is no evidence of an IC
bug, but there is obviously a problem.
I provided a patch that allows to disable the auto-seed, and in the
fsl_shw_get_random function, a seed is generated manually through a
descriptor every 16Mbytes. This number is higher than default auto-seed,
but was suggested by design team. There is no need to re-seed too ofter.
Alternative could be to do that every XX hours, or every YY requests to
that function, ...

The default schedule method to wait for events has been
replaced by "wait_event_interruptible_timeout" as in case
there is problem with SAHARA, such that issue, the driver
does not hang, and can report a problem.

Signed-off-by: Terry Lv <r65388@freescale.com>